Moved implementation of ConvectionEMP to ConvectionDiffusion
