+++ developers/UpwindQuadrature/mastersolution/upwindquadrature_main.cc
@@ -21,6 +21,7 @@
 #include <memory>
 
 #include "upwindquadrature.h"
+#include "../../../lecturecodes/ConvectionDiffusion/convection_emp.h"
 
 int main() {
   // PARAMETERS
@@ -88,7 +89,7 @@ int main() {
   lf::assemble::AssembleMatrixLocally(0, dofh, dofh, laplacian_provider, A);
 
   // Next part corresponding to the convection term:
-  UpwindQuadrature::ConvectionElementMatrixProvider convection_provider(v);
+  ConvectionDiffusion::ConvectionElementMatrixProvider convection_provider(v);
   lf::assemble::AssembleMatrixLocally(0, dofh, dofh, convection_provider, A);
 
   // RIGHT-HAND SIDE VECTOR
